% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distpower.r
\name{distPower}
\alias{distPower}
\alias{dpower}
\alias{ppower}
\alias{qpower}
\alias{rpower}
\title{Ditribution with a Power Law}
\usage{
dpower(x, pow, min = 0, max = 1)

ppower(x, pow, min = 0, max = 1)

qpower(p, pow, min = 0, max = 1)

rpower(n, pow, min = 0, max = 1)
}
\arguments{
\item{x}{vector of values where to evaluate the denisty or CDF.}

\item{pow}{degree of the power law.}

\item{min}{minimum value of the support of the distribution.}

\item{max}{maximum value of the support of the distribution.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If 'length(n) > 1', the length is taken to be the number reqired.}
}
\value{
‘dpower’ gives the density, ‘ppower’ gives the cumulative
distribution function (CDF), ‘qpower’ gives the quantile function
(i.e., the inverse of the CDF), and ‘rpower’ generates random numbers.

The length of the result is determined by ‘n’ for ‘rpower’, and is
the length of x or p for the other functions.
}
\description{
Density, distribution function, quantile function and random
generation for a continuous distribution with the density
(pow+1)*(x-min)^pow for x in \verb{[min,max]} and pow > -1.
}
