% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonboot.r
\name{estimate.tau}
\alias{estimate.tau}
\title{Estimating the convergence rate}
\usage{
estimate.tau(
  data,
  statistic,
  R = 1000,
  replace = FALSE,
  min.m = 3,
  beta = seq(0.2, 0.7, length.out = 5),
  method = "variance",
  ...
)
}
\arguments{
\item{data}{The data to be bootstrapped.}

\item{statistic}{The estimator of the parameter.}

\item{R}{amount of bootstrap replicates used to estimate tau.}

\item{replace}{If the sampling should be done with replacement.}

\item{min.m}{minimal subsampling size used to estimate tau.}

\item{beta}{the tested subsample sizes m are build by \code{n^beta}.}

\item{method}{the choosen method to estimate tau, choose one from \code{c("variance", "quantile")}.}

\item{...}{Additional parameters to be passed to the \code{mboot} function.}
}
\value{
A function for which applies \code{f(n) = tau_n}.
}
\description{
This function estimates the convergence rate of the bootstrap estimator
and returns it as a function of the form \code{tau_n = n^a}, where \code{n} is the input parameter.
}
\details{
There are two methods to choose from, \code{variance} and \code{politis}.
The provided \code{beta} values are used to select subsample sizes \code{m} by using \code{ms = n^beta}.
Note that the choice of the \code{beta} values can impact the accuracy of the estimated \code{tau}.
For each selected subsample size \code{m} a bootstrap with \code{R} replications is performed.
The method 'variance' then fits a linear function to the variance of the bootstrap statistics as function of the selected m values.
The method 'quantile' calculates multiple confidence intervals of different levels for each selected subsampling size and uses the difference
of these to estimatet the convergence rate.
}
\examples{
data <- runif(1000)
estimate.max <- function(data, indices) {return(max(data[indices]))}
estimated.tau <- estimate.tau(data, estimate.max)
boot.out <- mboot(data, estimate.max, R = 1000, replace = FALSE)
cis <- mboot.ci(boot.out, 0.95, estimated.tau, c("all"))
ci.basic <- cis$basic
print(ci.basic)

}
\references{
Bertail P. et al. (1999) On subsampling estimators with unknown rate of convergence.
\emph{Journal of the American Statistical Association}, 94(446):568-579.

Politis D.N. et al. (1999)
\emph{Subsampling}, Springer, New York.
}
\seealso{
mboot.ci
}
